Add support for disabling transmit
Add config support for disabling transmit via "-" in offset field.

If a single "-" is seen, the TX offset will be set to the max numeric
TX frequency – which is outside the valid freqency range. The TX
frequency, which is the offset summed with the RX frequency, is thus
also outside the valid range (as the RX frequency cannot be negative).

Modify setup_channel() to check whether the transmit freqency is in the
valid range. If not, rather than convert the TX frequency to BCD, hardcode
the TX frequency to 0xffffffff, the Baofeng's indicator for disabled TX.

See also: http://chirp.danplanet.com/issues/211

Move is_valid_frequency() to top of file to allow use from setup_channel().